pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-shield"
version = "0.1.0"
description = "MCP gateway with pattern-based security inspection — no LLM required"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
mcp-shield = "mcp_shield.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_shield"]
